Add the `--experimental-ssr` flag
