+++ examples/ssr/package.json
@@ -6,7 +6,7 @@
     "dev": "astro dev --experimental-ssr",
     "start": "astro dev",
     "build": "astro build --experimental-ssr",
-    "preview": "astro preview"
+    "server": "node server/server.mjs"
   },
   "devDependencies": {
     "astro": "^0.23.0-next.0",
