+++ examples/ssr/server/server.mjs
@@ -1,48 +1,15 @@
 import { createServer } from 'http';
-import { deserializeManifestData, matchRoute } from 'astro/routing';
-import { createResult } from 'astro/ssr/result.js';
-import { renderPage } from 'astro/server/index.js';
-import reactRenderer from '@astrojs/renderer-react';
-import reactServerRenderer from '@astrojs/renderer-react/server.js';
-import fs from 'fs';
-
-const renderers = [
-	Object.create(reactRenderer, {
-		ssr: {
-			value: reactServerRenderer
-		}
-	})
-];
+import { loadApp } from 'astro/app/node';
 
 const serverRoot = new URL('../dist/server/', import.meta.url);
-const manifestData = JSON.parse(await fs.promises.readFile(new URL('./manifest.json', serverRoot)));
-const manifest = deserializeManifestData(manifestData);
+const app = await loadApp(serverRoot);
 
 async function handle(req, res) {
-	const route = matchRoute(req.url, manifest);
+	const url = new URL(`http://${req.headers.host}${req.url}`);
+	const route = app.match(url);
+
 	if(route) {
-		const modFile = new URL('./' + route.distEntry, serverRoot);
-		const mod = await import(modFile);
-		const { default: Component } = mod;
-		// TODO get params
-
-		const result = createResult({
-			astroConfig: {
-				buildOptions: {}
-			},
-			logging: {},
-			origin: 'http://example.com',
-			params: {},
-			pathname: req.url,
-			renderers,
-			links: undefined, // TODO
-			scripts: undefined // TODO
-		});
-
-		const pageProps = {};
-
-		let html = await renderPage(result, Component, pageProps, null);
-
+		const html = await app.render(url, route);
 
 		res.writeHead(200, {
 			'Content-Type': 'text/html'
