+++ packages/astro/src/core/render/paginate.ts
