+++ packages/astro/src/core/render/script.ts
