+++ packages/astro/src/core/render/sitemap.ts
