+++ packages/astro/src/core/routing/index.ts
@@ -1,0 +1,11 @@
+export { createRouteManifest } from './manifest/create.js';
+export {
+	serializeRouteData,
+	deserializeRouteData
+} from './manifest/serialization.js';
+export { matchRoute } from './match.js';
+export { getParams } from './params.js';
+export {
+	validateGetStaticPathsModule,
+	validateGetStaticPathsResult
+} from './validation.js';
