+++ packages/astro/src/vite-plugin-astro-server/index.ts
@@ -2,17 +2,16 @@ import type vite from '../core/vite';
 import type http from 'http';
 import type { AstroConfig, ManifestData, RouteData } from '../@types/astro';
 import { info, LogOptions } from '../core/logger.js';
-import { fileURLToPath } from 'url';
-import { createRouteManifest, matchRoute } from '../core/ssr/routing.js';
+import { createRouteManifest, matchRoute } from '../core/routing/index.js';
 import mime from 'mime';
 import stripAnsi from 'strip-ansi';
 import { createSafeError } from '../core/util.js';
-import { ssr } from '../core/ssr/index.js';
+import { ssr } from '../core/render/dev/index.js';
 import * as msg from '../core/messages.js';
 
 import notFoundTemplate, { subpathNotUsedTemplate } from '../template/4xx.js';
 import serverErrorTemplate from '../template/5xx.js';
-import { RouteCache } from '../core/ssr/route-cache.js';
+import { RouteCache } from '../core/render/route-cache.js';
 
 interface AstroPluginOptions {
 	config: AstroConfig;
@@ -126,7 +125,6 @@ export default function createPlugin({ config, logging }: AstroPluginOptions): v
 	return {
 		name: 'astro:server',
 		configureServer(viteServer) {
-			const pagesDirectory = fileURLToPath(config.pages);
 			let routeCache = new RouteCache(logging);
 			let manifest: ManifestData = createRouteManifest({ config: config }, logging);
 			/** rebuild the route cache + manifest, as needed. */
